requirements.txt
Django==1.5.2
South==0.8
django-tastypie==0.10.0
mimeparse==0.1.3
Pillow==2.1.0

# Social login
python-openid
oauth2
django-social-auth
python-linkedin

# Score calculation
pyklout
python-twitter

# Requests
requests

# Worker
django-celery==3.0.23

